refactor(product): extract composer list-exit surgery into its own module
ComposerRichTextEditor.tsx crossed the 400-line frontend split threshold
(repo shape check); the PRO-267 list-exit helpers move to
ComposerListExit.tsx unchanged.
