+++ apps/packages/product-client/src/components/workspace/chat/input/ComposerRichTextEditor.tsx
@@ -1,21 +1,15 @@
 import { useEffect, useRef, type MutableRefObject, type Ref } from "react";
 import {
-  $createParagraphNode,
   $getRoot,
   $getSelection,
-  $isLineBreakNode,
   $isRangeSelection,
-  $isTextNode,
   COMMAND_PRIORITY_HIGH,
   INDENT_CONTENT_COMMAND,
   KEY_ENTER_COMMAND,
   KEY_TAB_COMMAND,
   OUTDENT_CONTENT_COMMAND,
   type LexicalEditor,
-  type LexicalNode,
-  type ParagraphNode,
 } from "lexical";
-import { $isListItemNode, $isListNode, type ListItemNode } from "@lexical/list";
 import {
   $convertFromMarkdownString,
   $convertToMarkdownString,
@@ -38,6 +32,11 @@ import {
   type ComposerEditorContext,
 } from "#product/components/workspace/chat/input/ComposerEditorDocument";
 import { ComposerLinkPastePlugin } from "#product/components/workspace/chat/input/ComposerLinkPastePlugin";
+import {
+  $exitListForShiftEnter,
+  $nearestListItem,
+  $outdentTopLevelListItems,
+} from "#product/components/workspace/chat/input/ComposerListExit";
 import { ComposerMarkdownShortcutPlugin } from "#product/components/workspace/chat/input/ComposerMarkdownShortcutPlugin";
 import { CHAT_TRANSCRIPT_LINK_CLASS } from "#product/config/transcript-link-styles";
 import type { ComposerKeyboardEventLike } from "#product/lib/domain/chat/composer/composer-keyboard";
@@ -361,110 +360,4 @@ function selectionIsInList(): boolean {
   const selection = $getSelection();
   if (!$isRangeSelection(selection)) return false;
   return $nearestListItem(selection.anchor.getNode()) !== null;
-}
-
-function $nearestListItem(node: LexicalNode): ListItemNode | null {
-  let current: LexicalNode | null = node;
-  while (current !== null && !$isListItemNode(current)) current = current.getParent();
-  return $isListItemNode(current) ? current : null;
-}
-
-// Lexical's stock OUTDENT_CONTENT_COMMAND only shrinks nesting, so it is a
-// silent no-op for items already at the top level (PRO-267). The missing
-// outdent step — item becomes a paragraph, splitting its list around it — is
-// owned here. Nested items (and wrapper items holding a nested list) still
-// take Lexical's own outdent path.
-function $outdentTopLevelListItems(): boolean {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection)) return false;
-  const items = new Set<ListItemNode>();
-  for (const node of selection.getNodes()) {
-    const item = $nearestListItem(node);
-    if (item) items.add(item);
-  }
-  if (items.size === 0) return false;
-  for (const item of items) {
-    if (item.getIndent() > 0 || item.getChildren().some($isListNode)) return false;
-  }
-  for (const item of items) $convertListItemToParagraph(item);
-  return true;
-}
-
-// Plain Enter submits the message, so Shift+Enter doubles as the list exit:
-// on an empty item it leaves the list (or un-nests one level), and on an
-// empty trailing line it moves the caret into a paragraph below the list.
-function $exitListForShiftEnter(editor: LexicalEditor): boolean {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection) || !selection.isCollapsed()) return false;
-  const item = $nearestListItem(selection.anchor.getNode());
-  if (!item) return false;
-  const effectivelyEmpty = item
-    .getChildren()
-    .every((child) => $isTextNode(child) && child.getTextContentSize() === 0);
-  if (effectivelyEmpty) {
-    if (item.getIndent() > 0) return editor.dispatchCommand(OUTDENT_CONTENT_COMMAND, undefined);
-    $convertListItemToParagraph(item);
-    return true;
-  }
-  if (item.getIndent() > 0) return false;
-  const { anchor } = selection;
-  const caretOnEmptyTrailingLine = $isLineBreakNode(item.getLastChild())
-    && anchor.type === "element"
-    && anchor.getNode().is(item)
-    && anchor.offset === item.getChildrenSize();
-  if (!caretOnEmptyTrailingLine) return false;
-  item.getLastChild()?.remove();
-  const list = item.getParentOrThrow();
-  const trailingStart = $orderedStartAfter(item);
-  const nextItem = item.getNextSibling();
-  const paragraph = $createParagraphNode();
-  item.insertAfter(paragraph);
-  $healTrailingList(paragraph, nextItem, trailingStart);
-  if (item.getChildrenSize() === 0) {
-    item.remove();
-    if (list.isAttached() && list.getChildrenSize() === 0) list.remove();
-  }
-  paragraph.selectEnd();
-  return true;
-}
-
-function $convertListItemToParagraph(item: ListItemNode): void {
-  const trailingStart = $orderedStartAfter(item);
-  const nextItem = item.getNextSibling();
-  // ListItemNode.replace splits the list around the item, transfers its
-  // children, remaps an element-anchored caret, and prunes the emptied list.
-  const paragraph = item.replace($createParagraphNode(), true);
-  $healTrailingList(paragraph, nextItem, trailingStart);
-}
-
-// Visible number the items after `item` should keep once a paragraph lands
-// before them, or null for unordered lists.
-function $orderedStartAfter(item: ListItemNode): number | null {
-  const list = item.getParent();
-  if (!$isListNode(list) || list.getListType() !== "number") return null;
-  return list.getStart() + item.getIndexWithinParent() + 1;
-}
-
-// After a paragraph lands between the halves of a split list, the trailing
-// half restarts ordered numbering at the head's start, and when the exited
-// item owned a nested sublist, that sublist's wrapper now leads the trailing
-// half as an indented bullet with no parent. Repair both.
-function $healTrailingList(
-  paragraph: ParagraphNode,
-  firstTrailingItem: LexicalNode | null,
-  orderedStart: number | null,
-): void {
-  if (!$isListItemNode(firstTrailingItem)) return;
-  const trailingList = firstTrailingItem.getParent();
-  if (!$isListNode(trailingList) || trailingList.getPreviousSibling() !== paragraph) return;
-  if (orderedStart !== null && trailingList.getListType() === "number") {
-    trailingList.setStart(orderedStart);
-  }
-  const first = trailingList.getFirstChild();
-  if (!$isListItemNode(first)) return;
-  const children = first.getChildren();
-  const orphanedSublist = children.length === 1 ? children[0] : null;
-  if (!$isListNode(orphanedSublist)) return;
-  for (const promoted of orphanedSublist.getChildren()) first.insertBefore(promoted);
-  first.remove();
 }
